+++ extras/AFC_utils.py
@@ -5,6 +5,20 @@
 # This file may be distributed under the terms of the GNU GPLv3 license.
 
 # File is used to hold common functions that can be called from anywhere and don't belong to a class
+import traceback
+import json
+
+from datetime import datetime
+from urllib.request import (
+    Request,
+    urlopen
+)
+from urllib.parse import (
+    urlencode,
+    urljoin
+)
+
+ERROR_STR = "Error trying to import {import_lib}, please rerun install-afc.sh script in your AFC-Klipper-Add-On directory then restart klipper\n\n{trace}"
 
 def add_filament_switch( switch_name, switch_pin, printer ):
     """
@@ -32,3 +46,167 @@ def add_filament_switch( switch_name, switch_pin, printer ):
     fila.runout_helper.runout_pause = False
 
     return fila
+
+def check_and_return( value_str:str, data_values:dict ) -> str:
+    """
+    Common function to check if value exists in dictionary and returns value if it does.
+
+    :param value_str: Key string to check if value exists in dictionary
+    :param data_values: Dictionary of values to check for key
+
+    :return: Returns string of value if found in dictionary
+    """
+    value = "0"
+    if value_str in data_values:
+        value = data_values[value_str]
+
+    return value
+
+class AFC_moonraker:
+    """
+    This class is used to communicate with moonraker to look up information and post
+    data into moonrakers database
+
+    Parameters
+    ----------------
+    port: String
+        Port to connect to moonrakers localhost
+    logger: AFC_logger
+        AFC logger object to log and print to console
+    """
+    ERROR_STRING = "Error getting data from moonraker, check AFC.log for more information"
+    def __init__(self, port:str, logger:object):
+        self.port           = port
+        self.logger         = logger
+        self.local_host     = 'http://localhost{port}'.format( port=port )
+        self.database_url   = urljoin(self.local_host, "server/database/item")
+        self.afc_stats_key  = "afc_stats"
+        self.afc_stats      = None
+        self.last_stats_time= None
+
+    def _get_results(self, url_string):
+        """
+        Helper function to get results, check for errors and return data if successful
+
+        :param url_string: URL encoded string to fetch/post data to moonraker
+
+        :returns: Returns result dictionary if data is valid , returns None if and error occurred
+        """
+        data = None
+        try:
+            resp = urlopen(url_string)
+            if resp.status >= 200 and resp.status <= 300:
+                data = json.load(resp)
+            else:
+                self.logger.error(self.ERROR_STRING)
+                self.logger.debug(f"Response: {resp.status} Reason: {resp.reason}")
+        except:
+            self.logger.error(self.ERROR_STRING, traceback=traceback.format_exc())
+            data = None
+        return data['result'] if data is not None else data
+
+    def get_spoolman_server(self)->str:
+        """
+        Queries moonraker to see if spoolman is configured, returns True when
+        spoolman is configured
+
+        :returns: Returns string for spoolmans IP, returns None if its not configured
+        """
+        resp = self._get_results(urljoin(self.local_host, 'server/config'))
+        # Check to make sure response is valid and spoolman exists in dictionary
+        if resp is not None and 'orig' in resp and 'spoolman' in resp['orig']:
+            return resp['orig']['spoolman']['server']     # check for spoolman and grab url
+        else:
+            self.logger.debug("Spoolman server is not defined")
+            return None
+
+    def get_file_filament_change_count(self, filename:str ):
+        """
+        Queries moonraker for files metadata and returns filament change count
+
+        :param filename: Filename to query moonraker and pull metadata
+        :return: Returns number of filament change counts if `filament_change_count` is in metadata.
+                 Returns zero if not found in metadata.
+        """
+        change_count = 0
+        resp = self._get_results(urljoin(self.local_host, 'server/files/metadata?filename={}'.format(filename)))
+        if resp is not None and 'filament_change_count' in resp:
+            change_count =  resp['filament_change_count']
+        else:
+            self.logger.debug(f"Metadata not found for file:{filename}")
+        return change_count
+
+    def get_afc_stats(self):
+        """
+        Queries moonraker database for all `afc_stats` entries and returns results if afc_stats exist.
+        Function also caches results and refetches data if cache is older than 60s. This is done to help
+        cut down on how much data is fetched from moonraker.
+
+        :return: Dictionary of afc_stats entries, None if afc_stats entry does not exist
+        """
+        resp = None
+        # Initially set to True since first time data always needs to be fetched
+        refetch_data = True
+        current_time = datetime.now()
+
+        # Check to see if data is older than 60 seconds and refreshes
+        if self.last_stats_time is not None:
+            refetch_data = False
+            delta = current_time - self.last_stats_time
+            if delta.seconds > 60:
+                refetch_data = True
+                self.last_stats_time = current_time
+        else:
+            self.last_stats_time = datetime.now()
+
+        # Cache results to keep queries to moonraker down
+        if self.afc_stats is None or refetch_data:
+            resp = self._get_results(urljoin(self.database_url, f"?namespace={self.afc_stats_key}"))
+            if resp is not None:
+                self.afc_stats = resp
+            else:
+                self.logger.debug("AFC_stats not in database")
+
+        return self.afc_stats
+
+    def update_afc_stats(self, key, value):
+        """
+        Updates afc_stats in moonrakers database with key, value pair
+
+        :param key: The key indicating the field where the value should be inserted
+        :param value: The value to insert into the database
+        """
+        resp = None
+        post_payload = {
+            "request_method": "POST",
+            "namespace": self.afc_stats_key,
+            "key": key,
+            "value": value
+        }
+        req = Request(self.database_url, urlencode(post_payload).encode())
+
+        resp = self._get_results(req)
+        if resp is None:
+            self.logger.error(f"Error when trying to update {key} in moonraker, see AFC.log for more info")
+
+    def get_spool(self, id:int):
+        """
+        Uses moonrakers proxy to query spoolID from spoolman
+
+        :param id: SpoolID to lookup and fetch data from spoolman
+        :return: Returns dictionary of spoolID, returns None if error occurred or ID does not exist
+        """
+        resp = None
+        request_payload = {
+            "request_method": "GET",
+            "path": f"/v1/spool/{id}"
+        }
+        spool_url = urljoin(self.local_host, 'server/spoolman/proxy')
+        req = Request( spool_url, urlencode(request_payload).encode() )
+
+        resp = self._get_results(req)
+        if resp is not None:
+            resp = resp
+        else:
+            self.logger.info(f"SpoolID: {id} not found")
+        return resp